+++ includes/minishell.h
@@ -17,7 +17,6 @@
 typedef struct s_str
 {
 	char			*str;
-	struct s_str	*prev;
 	struct s_str	*next;
 }t_str;
 
@@ -25,14 +24,16 @@ char	**modifsplit(char *input);
 void	dollarfound_getlen(int *a, int *count, char *input);
 void	expand_var(char *orgn, char **end, int *a, int *b);
 int		ft_modstrcpy(char *orgn, char **end, int len);
-void	interpreter(char **list);
+void	interpreter(char **list, t_str **env_list);
 void	ft_echo(char **list, int *idx);
 char	*adv_qm_rem(char *qm_str);
 int		qm_error_detector(char *str);
 void	init_ms_env(char **env_vector, t_str **env_list);
-void	print_env_list(char **env_list);
+void	ft_env(t_str *env_list);
 void	ft_pwd(void);
-void	ft_lstadd_back_str(t_str **list, char *str);
-void	ft_freelst_str(t_str *list);
-void	ft_printlist_str(t_str *list);
+void	ft_lst_str_add_back(t_str **list, char *str);
+void	ft_lst_str_free(t_str *list);
+void	ft_lst_str_print(t_str *list);
+void	ft_lst_str_add_front(t_str **list, char *str);
+void	ft_lst_str_delete(t_str **list, char *str);
 #endif