+++ src/utils/lst_str_handler.c
@@ -13,11 +13,32 @@
 #include "minishell.h"
 
 /**
+ * * Add new str into list at front
+ * @param list	list
+ * @param str	new str to link
+*/
+void	ft_lst_str_add_front(t_str **list, char *str)
+{
+	t_str	*new;
+
+	new = malloc(sizeof(t_str));
+	if (!new)
+	{
+		ft_lst_str_free(*list);
+		free(str);
+		exit(ENOMEM);
+	}
+	new->str = str;
+	new->next = *list;
+	*list = new;
+}
+
+/**
  * * Add new str into list at back
  * @param list	list
  * @param str	new str to link
 */
-void	ft_lstadd_back_str(t_str **list, char *str)
+void	ft_lst_str_add_back(t_str **list, char *str)
 {
 	t_str	*new;
 	t_str	*aux;
@@ -25,13 +46,12 @@ void	ft_lstadd_back_str(t_str **list, char *str)
 	new = malloc(sizeof(t_str));
 	if (!new)
 	{
-		ft_freelst_str(*list);
+		ft_lst_str_free(*list);
 		free(str);
 		exit(ENOMEM);
 	}
 	new->str = str;
 	new->next = NULL;
-	new->prev = NULL;
 	if (*list == NULL)
 		*list = new;
 	else
@@ -39,7 +59,6 @@ void	ft_lstadd_back_str(t_str **list, char *str)
 		aux = *list;
 		while (aux->next != NULL)
 			aux = aux->next;
-		new->prev = aux;
 		aux->next = new;
 	}
 }
@@ -48,7 +67,7 @@ void	ft_lstadd_back_str(t_str **list, char *str)
  * * Free list
  * @param list	list
 */
-void	ft_freelst_str(t_str *list)
+void	ft_lst_str_free(t_str *list)
 {
 	t_str	*next;
 	t_str	*aux;
@@ -68,7 +87,7 @@ void	ft_freelst_str(t_str *list)
  * * Print list
  * @param list	list
 */
-void	ft_printlist_str(t_str *list)
+void	ft_lst_str_print(t_str *list)
 {
 	while (list != NULL)
 	{
@@ -78,29 +97,23 @@ void	ft_printlist_str(t_str *list)
 }
 
 /**
- * * Delete str from list
+ * * Delete every str from list
  * @param list	list
  * @param str	str to delete
 */
-void	ft_lstdelete_str(t_str **list, char *str)
+void	ft_lst_str_delete(t_str **list, char *str)
 {
 	t_str	*prev;
 	t_str	*aux;
 	size_t	len;
 
 	len = ft_strlen(str);
-	while (ft_strcmp(*list->str, str, len))
-	{
-		aux = *list;
-		*list = *list->next;
-		free(aux->str);
-		free(aux);
-	}
-	prev = *listt;
+	ft_lst_str_add_front(list, NULL);
+	prev = *list;
+	aux = prev->next;
 	while (aux != NULL)
 	{
-		aux = prev->next;
-		if (ft_strcmp(aux->str, str, len))
+		if (!ft_strncmp(aux->str, str, len))
 		{
 			prev->next = aux->next;
 			free(aux->str);
@@ -108,6 +121,9 @@ void	ft_lstdelete_str(t_str **list, char *str)
 		}
 		else
 			prev = prev->next;
+		aux = prev->next;
 	}
+	aux = *list;
+	*list = aux->next;
+	free(aux);
 }
-