+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2021-11-03 - Edu
+
+### Added
+
+- **env.c** emulates bash pwd include functions to initialize env and shlvl (.src/buitins/env.c)
+- **lst_str_handlers.c** handle linked str list (.src/utils/lst_str_handlers.c)
+
+### Changed
+
+- **constants.h**addedd MAX_INT constant
+- Double linked str list to single linked list
+- **interpreter.c** ft_env added & env_list as function argument in some functions(command_finder & interpreter)
+
 ## 2021-11-01 - Edu
 
 ### Added
@@ -19,7 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- **Makefile** include missing header constants.h to fix not handling relink 
+- **Makefile** include missing header constants.h to fix not handling relink
 ## 2021-10-31 - Jorge
 
 ### Added